fix versioned tests to match original service test
not sure how I missed that existed already!
